Fix prettier errors and eslint warning about context
