+++ project-vite/eslint.config.cjs
@@ -1,50 +1,51 @@
-const {
-    defineConfig,
-    globalIgnores,
-} = require("eslint/config");
+const { defineConfig, globalIgnores } = require("eslint/config");
 
 const globals = require("globals");
 
-const {
-    fixupConfigRules,
-} = require("@eslint/compat");
+const { fixupConfigRules } = require("@eslint/compat");
 
 const tsParser = require("@typescript-eslint/parser");
 const reactRefresh = require("eslint-plugin-react-refresh");
 const js = require("@eslint/js");
 
-const {
-    FlatCompat,
-} = require("@eslint/eslintrc");
+const { FlatCompat } = require("@eslint/eslintrc");
 
 const compat = new FlatCompat({
     baseDirectory: __dirname,
     recommendedConfig: js.configs.recommended,
-    allConfig: js.configs.all
+    allConfig: js.configs.all,
 });
 
-module.exports = defineConfig([{
-    languageOptions: {
-        globals: {
-            ...globals.browser,
+module.exports = defineConfig([
+    {
+        languageOptions: {
+            globals: {
+                ...globals.browser,
+            },
+
+            parser: tsParser,
         },
 
-        parser: tsParser,
-    },
+        extends: fixupConfigRules(
+            compat.extends(
+                "eslint:recommended",
+                "plugin:@typescript-eslint/recommended",
+                "plugin:react-hooks/recommended"
+            )
+        ),
 
-    extends: fixupConfigRules(compat.extends(
-        "eslint:recommended",
-        "plugin:@typescript-eslint/recommended",
-        "plugin:react-hooks/recommended",
-    )),
+        plugins: {
+            "react-refresh": reactRefresh,
+        },
 
-    plugins: {
-        "react-refresh": reactRefresh,
+        rules: {
+            "react-refresh/only-export-components": [
+                "warn",
+                {
+                    allowConstantExport: true,
+                },
+            ],
+        },
     },
-
-    rules: {
-        "react-refresh/only-export-components": ["warn", {
-            allowConstantExport: true,
-        }],
-    },
-}, globalIgnores(["**/dist", "**/.eslintrc.cjs"])]);
+    globalIgnores(["**/dist", "**/.eslintrc.cjs"]),
+]);
